% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workload.R
\name{view_teaching_balance}
\alias{view_teaching_balance}
\title{View Teaching Balance}
\usage{
view_teaching_balance(workload)
}
\arguments{
\item{workload}{a dataframe returned from calculate_workload}
}
\value{
A grob object (a facted donut chart) visualizing the distribution
of teaching credits for each instructor.
}
\description{
Visualizes instructor teaching credits across the semester as a donut chart.
}
\examples{
# code to generate example workload dataframe here
# view_teaching_balance(2023_mds_workload)
}
