% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workload.R
\name{view_workload}
\alias{view_workload}
\title{View Instructor workload}
\usage{
view_workload(workload)
}
\arguments{
\item{workload}{a dataframe returned from calculate_workload}
}
\value{
a grob (ggplot2 object) that is a stacked bar chart
with a bar for each instructor and their workload proportions
mapped to colour
}
\description{
View Instructor workload
}
\examples{
# code to generate example workload dataframe here
view_workload(2023_mds_workload)
}
